tf

tf is a light-weight package with few dependencies that provides the class definitions and methods infrastructure for tidyfun -- tf gives you:

Details

new S3 data types for representing (vectors of) functional data: tfd() & tfb()

arithmetic operators for such data (Ops.tf()),

simple descriptive statistics: e.g. mean.tf(), median.tf()

base graphics functions for such data: plot.tf()

functions to do smoothing (tf_smooth.tfd()), differentiation tf_derive.tfd()) and integration (tf_derive.tfd())

tidyfun's goal is to make data wrangling and exploratory analysis for functional data in R quick and easy, using tidyverse syntax and standards.

Please install tidyfun for the full functionality to access the full documentation including a number of vignettes and case studies, or visit the tidyfun website.